refactor: extract bench side-by-side report
